refactor: add variant prop
