+++ app/_pages/SignIn.tsx
@@ -1,5 +1,5 @@
 import { PageWrapper } from '@/app/components/ui/PageWrapper';
-import { SignInForm } from '@/app/domains/auth/SignInForm';
+import { SignInForm } from '@/app/domains/auth/ui/SignInForm';
 
 export const SignInPage = () => (
   <PageWrapper>
