+++ app/_pages/SignUp.tsx
@@ -1,5 +1,5 @@
 import { PageWrapper } from '@/app/components/ui/PageWrapper';
-import { SignUpForm } from '@/app/domains/auth/SignUpForm';
+import { SignUpForm } from '@/app/domains/auth/ui/SignUpForm';
 
 export const SignUpPage = () => (
   <PageWrapper>
