+++ app/domains/auth/__tests__/SignInForm.spec.tsx
@@ -9,7 +9,7 @@ import {
   setupTranslationMocks,
 } from '@/app/__tests__/__mocks__/mock-setup';
 import { renderWithUserEvent, TestProviders } from '@/app/__tests__/utils';
-import { SignInForm } from '@/app/domains/auth/SignInForm';
+import { SignInForm } from '@/app/domains/auth/ui/SignInForm';
 
 describe('SignInForm', () => {
   let mockRegister: ReturnType<typeof vi.fn>;
