+++ app/domains/auth/ui/SignInForm.tsx
@@ -1,53 +1,50 @@
 'use client';
-import { Button, Fieldset, Stack } from '@chakra-ui/react';
+import { Button, Fieldset, Stack, Text } from '@chakra-ui/react';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useTranslations } from 'next-intl';
 import { useForm } from 'react-hook-form';
 import { FormField } from '@/app/components/ui/FormField';
-import { type SignUpFormType, signUpFormSchema } from '@/app/domains/auth/form-schemas';
+import { type SignInFormType, signInFormSchema } from '@/app/domains/auth/form-schemas';
 import { getValidationError } from '@/app/domains/auth/get-validation-error';
-import { PasswordField } from '@/app/domains/auth/PasswordField';
+import { PasswordField } from '@/app/domains/auth/ui/PasswordField';
 import { useAuth } from '@/app/hooks/use-auth';
 import { useAuthActions } from '@/app/hooks/use-auth-actions';
 import { useToast } from '@/app/hooks/use-toast';
+import { Link } from '@/i18n/routing';
 
-export const SignUpForm = () => {
+export const SignInForm = () => {
   const { success, error } = useToast();
+
   const t = useTranslations('form');
   const tNotification = useTranslations('notifications');
   const tValidation = useTranslations('validation');
 
-  const { signUp } = useAuthActions();
+  const { signIn } = useAuthActions();
   const { isLoading } = useAuth();
 
   const {
     register,
     handleSubmit,
     formState: { errors, isValid, isSubmitting },
-  } = useForm<SignUpFormType>({
+  } = useForm<SignInFormType>({
     mode: 'onChange',
     reValidateMode: 'onChange',
-    resolver: zodResolver(signUpFormSchema),
+    resolver: zodResolver(signInFormSchema),
   });
 
-  const handleSignUp = (data: SignUpFormType) =>
-    signUp(data)
-      .then(() => success(tNotification('signUpSuccess')))
+  const handleSignIn = (data: SignInFormType) =>
+    signIn(data)
+      .then(() => success(tNotification('signInSuccess')))
       .catch(() => error(tNotification('authError')));
 
   return (
-    <form onSubmit={handleSubmit(handleSignUp)}>
+    <form onSubmit={handleSubmit(handleSignIn)}>
       <Stack maxW="lg" w="full" mx="auto" p="8">
         <Fieldset.Root>
           <Fieldset.Legend fontSize="xl" fontWeight="bold">
-            {t('signUpTitle')}
+            {t('signInTitle')}
           </Fieldset.Legend>
         </Fieldset.Root>
-        <FormField
-          error={getValidationError(tValidation, errors.username?.message)}
-          label={t('username')}
-          {...register('username')}
-        />
         <FormField
           error={getValidationError(tValidation, errors.email?.message)}
           label={t('email')}
@@ -58,11 +55,6 @@ export const SignUpForm = () => {
           label={t('password')}
           {...register('password')}
         />
-        <PasswordField
-          error={getValidationError(tValidation, errors.confirmPassword?.message)}
-          label={t('confirmPassword')}
-          {...register('confirmPassword')}
-        />
         <Button
           data-testid="submit-button"
           loading={isLoading || isSubmitting}
@@ -72,6 +64,13 @@ export const SignUpForm = () => {
         >
           {t('submit')}
         </Button>
+
+        <Text textAlign="center" fontSize="md">
+          {t('noAccount')}{' '}
+          <Link className="!text-cyan-600 !font-bold" href="/sign-up">
+            {t('signUpHere')}
+          </Link>
+        </Text>
       </Stack>
     </form>
   );
